Hide redundant end points. Merge Variant and Variation. Default limit is 10 and default skip is 0. CamelCase instead of snake_case. #522
